Pull 'scrubData' helper function out as separate file
